Introduce the right Kan extension
Kan extensions are a way to take two functors `f` and `g`
and make a third that "extends" `f`s domain along `g`.
There are a few transformations from the canonical definition
of the right Kan extension to the definition here,
but this formulation should be equivalent.

Many people have said Kan extensions are the most universal constructions.
We show that here by defining Codensity and Yoneda in terms of Ran.