Add Arrows, and related classes and instances.
Mostly trying to get Category in, so we can generalize Functors, etc.

And this PR raises a few style questions:
1. Capitalization of types vs. their instances. In dada I’ve been using this
style, and so it’s in the directories I copied here. I prefer this style, but
not sure if there are objections to it.
2. Compose now has both a type and a class – is this a good naming convention?
Should we rename type to class where appropriate, or should these things be
renamed not to conflict or moved to separate directories or what?