Change order of arguments
As discussed in Slack,
it tends to be a bit easier to work with if the constraints are
pushed up as far as they can go.
We can partially apply ideas easier that way.

It seems like it depends on what the idea is,
since `Coyoneda/lift` improved at the expense of `Density/lift`.
On the whole, it seems easier.