Rethink `Ran/lift`
By thinking in terms of the universal property for the right Kan-extension,
dependents "stay on their side of the net."

For instance, `Codensity/lower` builds upon applicative stuff.
`Codensity/lift` can do similar, though it needs a full monad.
Instead of hopping the net to the comonad side.

Hopefully, this formulation is a bit clearer than the previous one.